Move DatapointKind back
