+++ tensorzero-core/tests/e2e/db/dataset_queries.rs
@@ -7,15 +7,13 @@ use tensorzero::{
     GetDatasetMetadataParams, Role,
 };
 use tensorzero_core::config::{MetricConfigLevel, MetricConfigType};
-use tensorzero_core::db::{
-    clickhouse::test_helpers::get_clickhouse,
-    datasets::{
-        ChatInferenceDatapointInsert, CountDatapointsForDatasetFunctionParams, DatapointInsert,
-        DatapointKind, DatasetMetadata, DatasetOutputSource, DatasetQueries,
-        GetAdjacentDatapointIdsParams, GetDatasetRowsParams, JsonInferenceDatapointInsert,
-        MetricFilter, StaleDatapointParams,
-    },
+use tensorzero_core::db::clickhouse::test_helpers::get_clickhouse;
+use tensorzero_core::db::datasets::{
+    ChatInferenceDatapointInsert, CountDatapointsForDatasetFunctionParams, DatapointInsert,
+    DatasetMetadata, DatasetOutputSource, DatasetQueries, GetAdjacentDatapointIdsParams,
+    GetDatasetRowsParams, JsonInferenceDatapointInsert, MetricFilter, StaleDatapointParams,
 };
+use tensorzero_core::endpoints::datasets::DatapointKind;
 use tensorzero_core::inference::types::{
     ContentBlockChatOutput, JsonInferenceOutput, StoredInput, StoredInputMessage,
     StoredInputMessageContent, Text,
